legacy txns bug fix and updated edit gas directories
