+++ app/components/UI/EditGasFee1559Swaps/index.js
@@ -1,6 +1,9 @@
+/* eslint-disable no-mixed-spaces-and-tabs */
+/* eslint-disable react/display-name */
 import React, { useCallback, useState, useMemo } from 'react';
 import {
   View,
+  StyleSheet,
   TouchableOpacity,
   ScrollView,
   TouchableWithoutFeedback,
@@ -14,7 +17,9 @@ import Icon from 'react-native-vector-icons/Ionicons';
 import { strings } from '../../../../locales/i18n';
 import Alert, { AlertType } from '../../Base/Alert';
 import HorizontalSelector from '../../Base/HorizontalSelector';
+import Device from '../../../util/device';
 import { isMainnetByChainId } from '../../../util/networks';
+import PropTypes from 'prop-types';
 import BigNumber from 'bignumber.js';
 import FadeAnimationView from '../FadeAnimationView';
 import { MetaMetricsEvents } from '../../../core/Analytics';
@@ -23,25 +28,144 @@ import AnalyticsV2 from '../../../util/analyticsV2';
 import TimeEstimateInfoModal from '../TimeEstimateInfoModal';
 import useModalHandler from '../../Base/hooks/useModalHandler';
 import AppConstants from '../../../core/AppConstants';
-import { useGasTransaction } from '../../../core/GasPolling/GasPolling';
-import { useAppThemeFromContext, mockTheme } from '../../../util/theme';
-import createStyles from './styles';
-import { EditGasFee1559UpdateProps, RenderInputProps } from './types';
+import { useTheme } from '../../../util/theme';
 
 const GAS_LIMIT_INCREMENT = new BigNumber(1000);
 const GAS_INCREMENT = new BigNumber(1);
 const GAS_LIMIT_MIN = new BigNumber(21000);
 const GAS_MIN = new BigNumber(0);
 
-const EditGasFee1559Update = ({
-  selectedGasValue,
+const createStyles = (colors) =>
+  StyleSheet.create({
+    root: {
+      backgroundColor: colors.background.default,
+      borderTopLeftRadius: 20,
+      borderTopRightRadius: 20,
+      minHeight: 200,
+      maxHeight: '95%',
+      paddingTop: 24,
+      paddingBottom: Device.isIphoneX() ? 32 : 24,
+    },
+    wrapper: {
+      paddingHorizontal: 24,
+    },
+    customGasHeader: {
+      flexDirection: 'row',
+      justifyContent: 'space-between',
+      alignItems: 'center',
+      width: '100%',
+      paddingBottom: 20,
+    },
+    newGasFeeHeader: {
+      flexDirection: 'row',
+      alignItems: 'center',
+      width: '100%',
+      justifyContent: 'center',
+    },
+    headerContainer: {
+      alignItems: 'center',
+      marginBottom: 22,
+    },
+    headerText: {
+      fontSize: 48,
+      flex: 1,
+      textAlign: 'center',
+    },
+    headerTitle: {
+      flexDirection: 'row',
+    },
+    saveButton: {
+      marginBottom: 20,
+    },
+    labelTextContainer: {
+      flexDirection: 'row',
+      alignItems: 'center',
+    },
+    hitSlop: {
+      top: 10,
+      left: 10,
+      bottom: 10,
+      right: 10,
+    },
+    labelInfo: {
+      color: colors.text.muted,
+    },
+    advancedOptionsContainer: {
+      marginTop: 25,
+      marginBottom: 30,
+    },
+    advancedOptionsInputsContainer: {
+      marginTop: 14,
+    },
+    rangeInputContainer: {
+      marginBottom: 20,
+    },
+    advancedOptionsButton: {
+      flexDirection: 'row',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+    advancedOptionsIcon: {
+      paddingTop: 1,
+      marginLeft: 5,
+    },
+    learnMoreLabels: {
+      marginTop: 9,
+    },
+    /* Add when the learn more link is ready
+  learnMoreLink: {
+    marginTop: 14
+  },*/
+    warningTextContainer: {
+      lineHeight: 20,
+      paddingLeft: 4,
+      flex: 1,
+    },
+    warningText: {
+      lineHeight: 20,
+      flex: 1,
+      color: colors.text.default,
+    },
+    warningContainer: {
+      marginBottom: 20,
+    },
+    dappEditGasContainer: {
+      marginVertical: 20,
+    },
+    subheader: {
+      marginBottom: 6,
+    },
+    learnMoreModal: {
+      maxHeight: Device.getDeviceHeight() * 0.7,
+    },
+    redInfo: {
+      marginLeft: 2,
+      color: colors.error.default,
+    },
+  });
+
+const EditGasFee1559Swaps = ({
+  selected,
+  gasFee,
   gasOptions,
+  onChange,
+  onCancel,
+  onSave,
+  gasFeeNative,
+  gasFeeConversion,
+  gasFeeMaxNative,
+  gasFeeMaxConversion,
+  maxPriorityFeeNative,
+  maxPriorityFeeConversion,
+  maxFeePerGasNative,
+  maxFeePerGasConversion,
   primaryCurrency,
   chainId,
-  onCancel,
-  onChange,
-  onSave,
+  timeEstimate,
+  timeEstimateColor,
+  timeEstimateId,
   error,
+  warning,
   dappSuggestedGas,
   ignoreOptions,
   updateOption,
@@ -51,75 +175,40 @@ const EditGasFee1559Update = ({
   suggestedEstimateOption,
   animateOnChange,
   isAnimating,
+  onUpdatingValuesStart,
+  onUpdatingValuesEnd,
   analyticsParams,
-  warning,
-  selectedGasObject,
-  onlyGas,
-}: EditGasFee1559UpdateProps) => {
-  const [modalInfo, updateModalInfo] = useState({
-    isVisible: false,
-    value: '',
-  });
-  const [showAdvancedOptions, setShowAdvancedOptions] = useState(
-    !selectedGasValue,
-  );
-  const [maxPriorityFeeError, setMaxPriorityFeeError] = useState('');
-  const [maxFeeError, setMaxFeeError] = useState('');
+  view,
+}) => {
+  const [showInfoModal, setShowInfoModal] = useState(false);
+  const [showAdvancedOptions, setShowAdvancedOptions] = useState(!selected);
+  const [maxPriorityFeeError, setMaxPriorityFeeError] = useState(null);
+  const [maxFeeError, setMaxFeeError] = useState(null);
   const [showLearnMoreModal, setShowLearnMoreModal] = useState(false);
-  const [selectedOption, setSelectedOption] = useState(selectedGasValue);
+  const [selectedOption, setSelectedOption] = useState(selected);
   const [showInputs, setShowInputs] = useState(!dappSuggestedGas);
-  const [gasObject, updateGasObject] = useState({
-    suggestedMaxFeePerGas: selectedGasObject.suggestedMaxFeePerGas,
-    suggestedMaxPriorityFeePerGas:
-      selectedGasObject.suggestedMaxPriorityFeePerGas,
-    suggestedGasLimit: selectedGasObject.suggestedGasLimit,
-  });
-
   const [
     isVisibleTimeEstimateInfoModal,
+    ,
     showTimeEstimateInfoModal,
     hideTimeEstimateInfoModal,
   ] = useModalHandler(false);
-  const { colors } = useAppThemeFromContext() || mockTheme;
+  const { colors } = useTheme();
   const styles = createStyles(colors);
-
-  const gasTransaction = useGasTransaction({
-    onlyGas,
-    gasSelected: selectedOption,
-    legacy: false,
-    gasObject,
-  });
-
-  const {
-    renderableGasFeeMinNative,
-    renderableGasFeeMaxNative,
-    renderableGasFeeMaxConversion,
-    renderableMaxFeePerGasNative,
-    renderableGasFeeMinConversion,
-    renderableMaxPriorityFeeNative,
-    renderableMaxFeePerGasConversion,
-    renderableMaxPriorityFeeConversion,
-    timeEstimateColor,
-    timeEstimate,
-    timeEstimateId,
-    suggestedMaxFeePerGas,
-    suggestedMaxPriorityFeePerGas,
-    suggestedGasLimit,
-  } = gasTransaction;
 
   const getAnalyticsParams = useCallback(() => {
     try {
       return {
         ...analyticsParams,
         chain_id: chainId,
-        function_type: analyticsParams.view,
+        function_type: view,
         gas_mode: selectedOption ? 'Basic' : 'Advanced',
         speed_set: selectedOption || undefined,
       };
-    } catch (err) {
+    } catch (error) {
       return {};
     }
-  }, [analyticsParams, chainId, selectedOption]);
+  }, [analyticsParams, chainId, selectedOption, view]);
 
   const toggleAdvancedOptions = useCallback(() => {
     if (!showAdvancedOptions) {
@@ -128,19 +217,12 @@ const EditGasFee1559Update = ({
         getAnalyticsParams(),
       );
     }
-    setShowAdvancedOptions(!showAdvancedOptions);
+    setShowAdvancedOptions((showAdvancedOptions) => !showAdvancedOptions);
   }, [getAnalyticsParams, showAdvancedOptions]);
 
   const toggleLearnMoreModal = useCallback(() => {
-    setShowLearnMoreModal(!showLearnMoreModal);
-  }, [showLearnMoreModal]);
-
-  const toggleInfoModal = useCallback(
-    (value: string) => {
-      updateModalInfo({ isVisible: !modalInfo.isVisible, value });
-    },
-    [updateModalInfo, modalInfo.isVisible],
-  );
+    setShowLearnMoreModal((showLearnMoreModal) => !showLearnMoreModal);
+  }, []);
 
   const save = useCallback(() => {
     AnalyticsV2.trackEvent(
@@ -148,35 +230,15 @@ const EditGasFee1559Update = ({
       getAnalyticsParams(),
     );
 
-    const newGasPriceObject = {
-      suggestedMaxFeePerGas: gasObject?.suggestedMaxFeePerGas,
-      suggestedMaxPriorityFeePerGas: gasObject?.suggestedMaxPriorityFeePerGas,
-      suggestedGasLimit: gasObject?.suggestedGasLimit,
-    };
-
-    onSave(gasTransaction, newGasPriceObject);
-  }, [getAnalyticsParams, onSave, gasTransaction, gasObject]);
+    onSave(selectedOption);
+  }, [getAnalyticsParams, onSave, selectedOption]);
 
   const changeGas = useCallback(
-    (gas, option) => {
-      setSelectedOption(option);
-      updateGasObject({
-        ...gasObject,
-        suggestedMaxFeePerGas: gas.suggestedMaxFeePerGas,
-        suggestedMaxPriorityFeePerGas: gas.suggestedMaxPriorityFeePerGas,
-        suggestedGasLimit: gas.suggestedGasLimit || gasObject.suggestedGasLimit,
-      });
-      onChange(option);
+    (gas, selectedOption) => {
+      setSelectedOption(selectedOption);
+      onChange(gas, selectedOption);
     },
-    [onChange, gasObject],
-  );
-
-  const changedGasLimit = useCallback(
-    (value) => {
-      const newGas = { ...gasTransaction, suggestedGasLimit: value };
-      changeGas(newGas, null);
-    },
-    [changeGas, gasTransaction],
+    [onChange],
   );
 
   const changedMaxPriorityFee = useCallback(
@@ -186,7 +248,6 @@ const EditGasFee1559Update = ({
           warningMinimumEstimateOption
         ]?.suggestedMaxPriorityFeePerGas,
       );
-
       const higherValue = new BigNumber(
         gasOptions?.high?.suggestedMaxPriorityFeePerGas,
       ).multipliedBy(new BigNumber(1.5));
@@ -213,23 +274,14 @@ const EditGasFee1559Update = ({
           strings('edit_gas_fee_eip1559.max_priority_fee_high'),
         );
       } else {
-        setMaxPriorityFeeError(null);
+        setMaxPriorityFeeError('');
       }
 
-      const newGas = {
-        ...gasTransaction,
-        suggestedMaxPriorityFeePerGas: value,
-      };
+      const newGas = { ...gasFee, suggestedMaxPriorityFeePerGas: value };
 
       changeGas(newGas, null);
     },
-    [
-      changeGas,
-      gasTransaction,
-      gasOptions,
-      updateOption,
-      warningMinimumEstimateOption,
-    ],
+    [changeGas, gasFee, gasOptions, updateOption, warningMinimumEstimateOption],
   );
 
   const changedMaxFeePerGas = useCallback(
@@ -262,20 +314,18 @@ const EditGasFee1559Update = ({
         setMaxFeeError('');
       }
 
-      const newGas = {
-        ...gasTransaction,
-        suggestedMaxFeePerGas: value,
-      };
-
+      const newGas = { ...gasFee, suggestedMaxFeePerGas: value };
       changeGas(newGas, null);
     },
-    [
-      changeGas,
-      gasTransaction,
-      gasOptions,
-      updateOption,
-      warningMinimumEstimateOption,
-    ],
+    [changeGas, gasFee, gasOptions, updateOption, warningMinimumEstimateOption],
+  );
+
+  const changedGasLimit = useCallback(
+    (value) => {
+      const newGas = { ...gasFee, suggestedGasLimit: value };
+      changeGas(newGas, null);
+    },
+    [changeGas, gasFee],
   );
 
   const selectOption = useCallback(
@@ -283,13 +333,13 @@ const EditGasFee1559Update = ({
       setSelectedOption(option);
       setMaxFeeError('');
       setMaxPriorityFeeError('');
-      changeGas({ ...gasOptions?.[option] }, option);
+      changeGas({ ...gasOptions[option] }, option);
     },
     [changeGas, gasOptions],
   );
 
   const shouldIgnore = useCallback(
-    (option) => ignoreOptions?.find((item: string) => item === option),
+    (option) => ignoreOptions.find((item) => item === option),
     [ignoreOptions],
   );
 
@@ -312,13 +362,11 @@ const EditGasFee1559Update = ({
         .filter(({ name }) => !shouldIgnore(name))
         .map(({ name, label, ...option }) => ({
           name,
-          label: function LabelComponent(selected: any, disabled: any) {
-            return (
-              <Text bold primary={selected && !disabled}>
-                {label}
-              </Text>
-            );
-          },
+          label: (selected, disabled) => (
+            <Text bold primary={selected && !disabled}>
+              {label}
+            </Text>
+          ),
           topLabel: recommended?.name === name && recommended.render,
           ...option,
           ...extendOptions[name],
@@ -328,74 +376,37 @@ const EditGasFee1559Update = ({
 
   const isMainnet = isMainnetByChainId(chainId);
   const nativeCurrencySelected = primaryCurrency === 'ETH' || !isMainnet;
+  let gasFeePrimary,
+    gasFeeMaxPrimary,
+    maxFeePerGasPrimary,
+    maxPriorityFeePerGasPrimary,
+    gasFeeMaxSecondary;
+  if (nativeCurrencySelected) {
+    gasFeePrimary = gasFeeNative;
+    gasFeeMaxPrimary = gasFeeMaxNative;
+    gasFeeMaxSecondary = gasFeeMaxConversion;
+    maxFeePerGasPrimary = maxFeePerGasNative;
+    maxPriorityFeePerGasPrimary = maxPriorityFeeNative;
+  } else {
+    gasFeePrimary = gasFeeConversion;
+    gasFeeMaxPrimary = gasFeeMaxConversion;
+    gasFeeMaxSecondary = gasFeeMaxNative;
+    maxFeePerGasPrimary = maxFeePerGasConversion;
+    maxPriorityFeePerGasPrimary = maxPriorityFeeConversion;
+  }
 
-  const switchNativeCurrencyDisplayOptions = (
-    nativeValue: string,
-    fiatValue: string,
-  ) => {
-    if (nativeCurrencySelected) return nativeValue;
-    return fiatValue;
-  };
+  const valueToWatch = `${gasFeeNative}${gasFeeMaxNative}`;
 
-  const valueToWatch = `${renderableGasFeeMinNative}${renderableGasFeeMaxNative}`;
-
-  const LeftLabelComponent = ({
-    value,
-    infoValue,
-  }: {
-    value: string;
-    infoValue: string;
-  }) => (
-    <View style={styles.labelTextContainer}>
-      <Text black bold noMargin>
-        {strings(value)}
-      </Text>
-      <TouchableOpacity
-        hitSlop={styles.hitSlop}
-        onPress={() => toggleInfoModal(infoValue)}
-      >
-        <MaterialCommunityIcon
-          name="information"
-          size={14}
-          style={styles.labelInfo}
-        />
-      </TouchableOpacity>
-    </View>
-  );
-
-  const RightLabelComponent = ({ value }: { value: string }) => (
-    <Text noMargin small grey>
-      <Text bold reset>
-        {strings(value)}:
-      </Text>{' '}
-      {gasOptions?.[suggestedEstimateOption]?.suggestedMaxFeePerGas} GWEI
-    </Text>
-  );
-
-  const TextComponent = ({
-    title,
-    value,
-  }: {
-    title: string;
-    value: string;
-  }) => (
-    <>
-      <Text noMargin primary infoModal bold style={styles.learnMoreLabels}>
-        {strings(title)}
-      </Text>
-      <Text noMargin grey infoModal>
-        {strings(value)}
-      </Text>
-    </>
-  );
-
-  const renderInputs = (option: RenderInputProps) => (
+  const renderInputs = () => (
     <View>
       <FadeAnimationView
         valueToWatch={valueToWatch}
         animateOnChange={animateOnChange}
+        onAnimationStart={onUpdatingValuesStart}
+        onAnimationEnd={onUpdatingValuesEnd}
       >
         <View>
+          {/* TODO(eip1559) hook with strings i18n */}
           <HorizontalSelector
             selected={selectedOption}
             onPress={selectOption}
@@ -404,7 +415,7 @@ const EditGasFee1559Update = ({
         </View>
         <View style={styles.advancedOptionsContainer}>
           <TouchableOpacity
-            disabled={option?.showAdvanced}
+            disable={updateOption?.showAdvanced}
             onPress={toggleAdvancedOptions}
             style={styles.advancedOptionsButton}
           >
@@ -415,18 +426,30 @@ const EditGasFee1559Update = ({
               <Icon name={`ios-arrow-${showAdvancedOptions ? 'up' : 'down'}`} />
             </Text>
           </TouchableOpacity>
-          {(showAdvancedOptions || option?.maxFeeThreshold) && (
+          {(showAdvancedOptions || updateOption?.showAdvanced) && (
             <View style={styles.advancedOptionsInputsContainer}>
               <View style={styles.rangeInputContainer}>
                 <RangeInput
                   leftLabelComponent={
-                    <LeftLabelComponent
-                      value="edit_gas_fee_eip1559.gas_limit"
-                      infoValue="gas_limit"
-                    />
+                    <View style={styles.labelTextContainer}>
+                      <Text black bold noMargin>
+                        {strings('edit_gas_fee_eip1559.gas_limit')}{' '}
+                      </Text>
+
+                      <TouchableOpacity
+                        hitSlop={styles.hitSlop}
+                        onPress={() => setShowInfoModal('gas_limit')}
+                      >
+                        <MaterialCommunityIcon
+                          name="information"
+                          size={14}
+                          style={styles.labelInfo}
+                        />
+                      </TouchableOpacity>
+                    </View>
                   }
                   min={GAS_LIMIT_MIN}
-                  value={suggestedGasLimit}
+                  value={gasFee.suggestedGasLimit}
                   onChangeValue={changedGasLimit}
                   name={strings('edit_gas_fee_eip1559.gas_limit')}
                   increment={GAS_LIMIT_INCREMENT}
@@ -435,25 +458,43 @@ const EditGasFee1559Update = ({
               <View style={styles.rangeInputContainer}>
                 <RangeInput
                   leftLabelComponent={
-                    <LeftLabelComponent
-                      value="edit_gas_fee_eip1559.max_priority_fee"
-                      infoValue="max_priority_fee"
-                    />
+                    <View style={styles.labelTextContainer}>
+                      <Text black bold noMargin>
+                        {strings('edit_gas_fee_eip1559.max_priority_fee')}{' '}
+                      </Text>
+
+                      <TouchableOpacity
+                        hitSlop={styles.hitSlop}
+                        onPress={() => setShowInfoModal('max_priority_fee')}
+                      >
+                        <MaterialCommunityIcon
+                          name="information"
+                          size={14}
+                          style={styles.labelInfo}
+                        />
+                      </TouchableOpacity>
+                    </View>
                   }
                   rightLabelComponent={
-                    <RightLabelComponent value="edit_gas_fee_eip1559.estimate" />
+                    <Text noMargin small grey>
+                      <Text bold reset>
+                        {strings('edit_gas_fee_eip1559.estimate')}:
+                      </Text>{' '}
+                      {
+                        gasOptions?.[suggestedEstimateOption]
+                          ?.suggestedMaxPriorityFeePerGas
+                      }{' '}
+                      GWEI
+                    </Text>
                   }
-                  value={suggestedMaxPriorityFeePerGas}
+                  value={gasFee.suggestedMaxPriorityFeePerGas}
                   name={strings('edit_gas_fee_eip1559.max_priority_fee')}
                   unit={'GWEI'}
                   min={GAS_MIN}
                   increment={GAS_INCREMENT}
                   inputInsideLabel={
-                    renderableMaxPriorityFeeNative &&
-                    `≈ ${switchNativeCurrencyDisplayOptions(
-                      renderableMaxPriorityFeeNative,
-                      renderableMaxPriorityFeeConversion,
-                    )}`
+                    maxPriorityFeePerGasPrimary &&
+                    `≈ ${maxPriorityFeePerGasPrimary}`
                   }
                   error={maxPriorityFeeError}
                   onChangeValue={changedMaxPriorityFee}
@@ -462,15 +503,41 @@ const EditGasFee1559Update = ({
               <View style={styles.rangeInputContainer}>
                 <RangeInput
                   leftLabelComponent={
-                    <LeftLabelComponent
-                      value="edit_gas_fee_eip1559.max_fee"
-                      infoValue="max_fee"
-                    />
+                    <View style={styles.labelTextContainer}>
+                      <Text
+                        black={!maxFeeError}
+                        red={Boolean(maxFeeError)}
+                        bold
+                        noMargin
+                      >
+                        {strings('edit_gas_fee_eip1559.max_fee')}{' '}
+                      </Text>
+
+                      <TouchableOpacity
+                        hitSlop={styles.hitSlop}
+                        onPress={() => setShowInfoModal('max_fee')}
+                      >
+                        <MaterialCommunityIcon
+                          name="information"
+                          size={14}
+                          style={styles.labelInfo}
+                        />
+                      </TouchableOpacity>
+                    </View>
                   }
                   rightLabelComponent={
-                    <RightLabelComponent value="edit_gas_fee_eip1559.estimate" />
+                    <Text noMargin small grey>
+                      <Text bold reset>
+                        {strings('edit_gas_fee_eip1559.estimate')}:
+                      </Text>{' '}
+                      {
+                        gasOptions?.[suggestedEstimateOption]
+                          ?.suggestedMaxFeePerGas
+                      }{' '}
+                      GWEI
+                    </Text>
                   }
-                  value={suggestedMaxFeePerGas}
+                  value={gasFee.suggestedMaxFeePerGas}
                   name={strings('edit_gas_fee_eip1559.max_fee')}
                   unit={'GWEI'}
                   min={GAS_MIN}
@@ -478,11 +545,7 @@ const EditGasFee1559Update = ({
                   error={maxFeeError}
                   onChangeValue={changedMaxFeePerGas}
                   inputInsideLabel={
-                    renderableMaxFeePerGasNative &&
-                    `≈ ${switchNativeCurrencyDisplayOptions(
-                      renderableMaxFeePerGasNative,
-                      renderableMaxFeePerGasConversion,
-                    )}`
+                    maxFeePerGasPrimary && `≈ ${maxFeePerGasPrimary}`
                   }
                 />
               </View>
@@ -504,7 +567,7 @@ const EditGasFee1559Update = ({
           onPress={save}
           disabled={Boolean(error) || isAnimating}
         >
-          {option
+          {updateOption
             ? strings('edit_gas_fee_eip1559.submit')
             : strings('edit_gas_fee_eip1559.save')}
         </StyledButton>
@@ -609,7 +672,7 @@ const EditGasFee1559Update = ({
 
                   <TouchableOpacity
                     hitSlop={styles.hitSlop}
-                    onPress={() => toggleInfoModal('new_gas_fee')}
+                    onPress={() => setShowInfoModal('new_gas_fee')}
                   >
                     <MaterialCommunityIcon
                       name="information"
@@ -635,27 +698,14 @@ const EditGasFee1559Update = ({
                   numberOfLines={1}
                   noMargin
                 >
-                  ~
-                  {switchNativeCurrencyDisplayOptions(
-                    renderableGasFeeMinNative,
-                    renderableGasFeeMinConversion,
-                  )}
+                  ~{gasFeePrimary}
                 </Text>
               </View>
               <Text big black style={styles.subheader} noMargin>
                 <Text bold black noMargin>
                   {strings('edit_gas_fee_eip1559.max_fee')}:{' '}
                 </Text>
-                {switchNativeCurrencyDisplayOptions(
-                  renderableGasFeeMaxNative,
-                  renderableGasFeeMaxConversion,
-                )}{' '}
-                (
-                {switchNativeCurrencyDisplayOptions(
-                  renderableGasFeeMaxConversion,
-                  renderableGasFeeMaxNative,
-                )}
-                )
+                {gasFeeMaxPrimary} ({gasFeeMaxSecondary})
               </Text>
               <View style={styles.labelTextContainer}>
                 <Text
@@ -665,12 +715,11 @@ const EditGasFee1559Update = ({
                 >
                   {timeEstimate}
                 </Text>
-                {timeEstimateId ===
-                  (AppConstants.GAS_TIMES.MAYBE ||
-                    AppConstants.GAS_TIMES.UNKNOWN) && (
+                {(timeEstimateId === AppConstants.GAS_TIMES.MAYBE ||
+                  timeEstimateId === AppConstants.GAS_TIMES.UNKNOWN) && (
                   <TouchableOpacity
                     hitSlop={styles.hitSlop}
-                    onPress={() => showTimeEstimateInfoModal()}
+                    onPress={showTimeEstimateInfoModal}
                   >
                     <MaterialCommunityIcon
                       name="information"
@@ -691,37 +740,34 @@ const EditGasFee1559Update = ({
                 </StyledButton>
               </View>
             ) : (
-              renderInputs(updateOption)
+              renderInputs()
             )}
-
             <InfoModal
-              isVisible={Boolean(modalInfo.isVisible)}
+              isVisible={Boolean(showInfoModal)}
               title={
-                modalInfo.value === 'gas_limit'
+                showInfoModal === 'gas_limit'
                   ? strings('edit_gas_fee_eip1559.gas_limit')
-                  : modalInfo.value === 'max_priority_fee'
+                  : showInfoModal === 'max_priority_fee'
                   ? strings('edit_gas_fee_eip1559.max_priority_fee')
-                  : modalInfo.value === 'max_fee'
+                  : showInfoModal === 'max_fee'
                   ? strings('edit_gas_fee_eip1559.max_fee')
-                  : modalInfo.value === 'new_gas_fee'
+                  : showInfoModal === 'new_gas_fee'
                   ? strings('edit_gas_fee_eip1559.new_gas_fee')
                   : null
               }
-              toggleModal={() =>
-                updateModalInfo({ ...modalInfo, isVisible: false })
-              }
+              toggleModal={() => setShowInfoModal(null)}
               body={
                 <View>
                   <Text grey infoModal>
-                    {modalInfo.value === 'gas_limit' &&
+                    {showInfoModal === 'gas_limit' &&
                       strings('edit_gas_fee_eip1559.learn_more_gas_limit')}
-                    {modalInfo.value === 'max_priority_fee' &&
+                    {showInfoModal === 'max_priority_fee' &&
                       strings(
                         'edit_gas_fee_eip1559.learn_more_max_priority_fee',
                       )}
-                    {modalInfo.value === 'max_fee' &&
+                    {showInfoModal === 'max_fee' &&
                       strings('edit_gas_fee_eip1559.learn_more_max_fee')}
-                    {modalInfo.value === 'new_gas_fee' &&
+                    {showInfoModal === 'new_gas_fee' &&
                     updateOption &&
                     updateOption.isCancel
                       ? strings(
@@ -745,18 +791,56 @@ const EditGasFee1559Update = ({
                         <Text noMargin grey infoModal>
                           {strings('edit_gas_fee_eip1559.learn_more.intro')}
                         </Text>
-                        <TextComponent
-                          title="edit_gas_fee_eip1559.learn_more.low_label"
-                          value="edit_gas_fee_eip1559.learn_more.low_text"
-                        />
-                        <TextComponent
-                          title="edit_gas_fee_eip1559.learn_more.market_label"
-                          value="edit_gas_fee_eip1559.learn_more.market_text"
-                        />
-                        <TextComponent
-                          title="edit_gas_fee_eip1559.learn_more.aggressive_label"
-                          value="edit_gas_fee_eip1559.learn_more.aggressive_text"
-                        />
+                        <Text
+                          noMargin
+                          primary
+                          infoModal
+                          bold
+                          style={styles.learnMoreLabels}
+                        >
+                          {strings('edit_gas_fee_eip1559.learn_more.low_label')}
+                        </Text>
+                        <Text noMargin grey infoModal>
+                          {strings('edit_gas_fee_eip1559.learn_more.low_text')}
+                        </Text>
+                        <Text
+                          noMargin
+                          primary
+                          infoModal
+                          bold
+                          style={styles.learnMoreLabels}
+                        >
+                          {strings(
+                            'edit_gas_fee_eip1559.learn_more.market_label',
+                          )}
+                        </Text>
+                        <Text noMargin grey infoModal>
+                          {strings(
+                            'edit_gas_fee_eip1559.learn_more.market_text',
+                          )}
+                        </Text>
+                        <Text
+                          noMargin
+                          primary
+                          infoModal
+                          bold
+                          style={styles.learnMoreLabels}
+                        >
+                          {strings(
+                            'edit_gas_fee_eip1559.learn_more.aggressive_label',
+                          )}
+                        </Text>
+                        <Text noMargin grey infoModal>
+                          {strings(
+                            'edit_gas_fee_eip1559.learn_more.aggressive_text',
+                          )}
+                        </Text>
+                        {/* TODO(eip1559) add link when available
+                        <TouchableOpacity style={styles.learnMoreLink}>
+                          <Text grey infoModal link>
+                            {strings('edit_gas_fee_eip1559.learn_more.link')}
+                          </Text>
+                        </TouchableOpacity>*/}
                       </View>
                     </TouchableWithoutFeedback>
                   </ScrollView>
@@ -775,4 +859,157 @@ const EditGasFee1559Update = ({
   );
 };
 
-export default EditGasFee1559Update;
+EditGasFee1559Swaps.defaultProps = {
+  ignoreOptions: [],
+  warningMinimumEstimateOption: AppConstants.GAS_OPTIONS.LOW,
+  suggestedEstimateOption: AppConstants.GAS_OPTIONS.MEDIUM,
+};
+
+EditGasFee1559Swaps.propTypes = {
+  /**
+   * Gas option selected (low, medium, high)
+   */
+  selected: PropTypes.string,
+  /**
+   * Gas fee currently active
+   */
+  gasFee: PropTypes.object,
+  /**
+   * Gas fee options to select from
+   */
+  gasOptions: PropTypes.object,
+  /**
+   * Function called when user selected or changed the gas
+   */
+  onChange: PropTypes.func,
+  /**
+   * Function called when user cancels
+   */
+  onCancel: PropTypes.func,
+  /**
+   * Function called when user saves the new gas
+   */
+  onSave: PropTypes.func,
+  /**
+   * Gas fee in native currency
+   */
+  gasFeeNative: PropTypes.string,
+  /**
+   * Gas fee converted to chosen currency
+   */
+  gasFeeConversion: PropTypes.string,
+  /**
+   * Maximum gas fee in native currency
+   */
+  gasFeeMaxNative: PropTypes.string,
+  /**
+   * Maximum gas fee converted to chosen currency
+   */
+  gasFeeMaxConversion: PropTypes.string,
+  /**
+   * Maximum priority gas fee in native currency
+   */
+  maxPriorityFeeNative: PropTypes.string,
+  /**
+   * Maximum priority gas fee converted to chosen currency
+   */
+  maxPriorityFeeConversion: PropTypes.string,
+  /**
+   * Maximum fee per gas fee in native currency
+   */
+  maxFeePerGasNative: PropTypes.string,
+  /**
+   * Maximum fee per gas fee converted to chosen currency
+   */
+  maxFeePerGasConversion: PropTypes.string,
+  /**
+   * Primary currency, either ETH or Fiat
+   */
+  primaryCurrency: PropTypes.string,
+  /**
+   * A string representing the network chainId
+   */
+  chainId: PropTypes.string,
+  /**
+   * String that represents the time estimates
+   */
+  timeEstimate: PropTypes.string,
+  /**
+   * String that represents the color of the time estimate
+   */
+  timeEstimateColor: PropTypes.string,
+  /**
+   * Time estimate name (unknown, low, medium, high, less_than, range)
+   */
+  timeEstimateId: PropTypes.string,
+  /**
+   * Error message to show
+   */
+  error: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.bool,
+    PropTypes.node,
+  ]),
+  /**
+   * Warning message to show
+   */
+  warning: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.bool,
+    PropTypes.node,
+  ]),
+  /**
+   * Boolean that specifies if the gas price was suggested by the dapp
+   */
+  dappSuggestedGas: PropTypes.bool,
+  /**
+   * Ignore option array
+   */
+  ignoreOptions: PropTypes.array,
+  /**
+   * Option to display speed up/cancel view
+   */
+  updateOption: PropTypes.object,
+  /**
+   * Extend options object. Object has option keys and properties will be spread
+   */
+  extendOptions: PropTypes.object,
+  /**
+   * Recommended object with type and render function
+   */
+  recommended: PropTypes.object,
+  /**
+   * Estimate option to compare with for too low warning
+   */
+  warningMinimumEstimateOption: PropTypes.string,
+  /**
+   * Suggested estimate option to show recommended values
+   */
+  suggestedEstimateOption: PropTypes.string,
+  /**
+   * Function to call when update animation starts
+   */
+  onUpdatingValuesStart: PropTypes.func,
+  /**
+   * Function to call when update animation ends
+   */
+  onUpdatingValuesEnd: PropTypes.func,
+  /**
+   * If the values should animate upon update or not
+   */
+  animateOnChange: PropTypes.bool,
+  /**
+   * Boolean to determine if the animation is happening
+   */
+  isAnimating: PropTypes.bool,
+  /**
+   * Extra analytics params to be send with the gas analytics
+   */
+  analyticsParams: PropTypes.object,
+  /**
+   * (For analytics purposes) View (Approve, Transfer, Confirm) where this component is being used
+   */
+  view: PropTypes.string.isRequired,
+};
+
+export default EditGasFee1559Swaps;
