+++ app/components/UI/EditGasFee1559Swaps/index.test.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import EditGasFee1559 from './';
+import EditGasFee1559Swaps from '.';
 
-describe('EditGasFee1559', () => {
+describe('EditGasFee1559Swaps', () => {
   it('should render correctly', () => {
     const wrapper = shallow(
-      <EditGasFee1559
+      <EditGasFee1559Swaps
         gasFee={{
           maxWaitTimeEstimate: 150000,
           minWaitTimeEstimate: 0,
