+++ app/components/UI/TransactionEditor/index.js
@@ -24,8 +24,8 @@ import Engine from '../../../core/Engine';
 import collectiblesTransferInformation from '../../../util/collectibles-transfer';
 import { safeToChecksumAddress } from '../../../util/address';
 import { shallowEqual } from '../../../util/general';
-import EditGasFee1559 from '../EditGasFee1559Update';
-import EditGasFeeLegacy from '../EditGasFeeLegacyUpdate';
+import EditGasFee1559 from '../EditGasFee1559';
+import EditGasFeeLegacy from '../EditGasFeeLegacy';
 import { GAS_ESTIMATE_TYPES } from '@metamask/gas-fee-controller';
 import AppConstants from '../../../core/AppConstants';
 import {
