+++ app/components/UI/TransactionReview/TransactionReviewInformation/index.js
@@ -591,6 +591,7 @@ class TransactionReviewInformation extends PureComponent {
       isAnimating,
       multiLayerL1FeeTotal,
       gasObject,
+      gasSelected,
       updateTransactionState,
       dappSuggestedEIP1559Gas,
       dappSuggestedGasPrice,
@@ -627,6 +628,7 @@ class TransactionReviewInformation extends PureComponent {
         dappSuggestedEIP1559Gas={dappSuggestedEIP1559Gas}
         updateTransactionState={updateTransactionState}
         onlyGas
+        gasSelected={gasSelected}
         dappSuggestedGasPrice={dappSuggestedGasPrice}
       />
     );
