+++ app/components/UI/UpdateEIP1559Tx/index.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable no-mixed-spaces-and-tabs */
 import React, { useState, useEffect, useCallback, useRef } from 'react';
-import EditGasFee1559Update from '../EditGasFee1559Update';
+import EditGasFee1559Update from '../EditGasFee1559';
 import { connect } from 'react-redux';
 import { CANCEL_RATE, SPEED_UP_RATE } from '@metamask/transaction-controller';
 import { GAS_ESTIMATE_TYPES } from '@metamask/gas-fee-controller';
